+++ controllers/datadogagent/clusteragent.go
@@ -1319,14 +1319,6 @@ func getExternalMetricsEndpoint(dda *datadoghqv1alpha1.DatadogAgent) string {
 	return "https://app.datadoghq.com"
 }
 
-func (r *Reconciler) manageKubeStateMetricsCore(logger logr.Logger, dda *datadoghqv1alpha1.DatadogAgent) (reconcile.Result, error) {
-	if !isKSMCoreEnabled(dda.Spec.ClusterAgent) {
-		return reconcile.Result{}, nil
-	}
-	// Only create the default ConfigMap if the conf is not overridden
-	return r.manageConfigMap(logger, dda, datadoghqv1alpha1.GetKubeStateMetricsConfName(dda), buildKSMCoreConfigMap)
-}
-
 func (r *Reconciler) manageClusterAgentNetworkPolicy(logger logr.Logger, dda *datadoghqv1alpha1.DatadogAgent) (reconcile.Result, error) {
 	policyName := fmt.Sprintf("%s-%s", dda.Name, datadoghqv1alpha1.DefaultClusterAgentResourceSuffix)
 
@@ -1436,262 +1428,4 @@ func buildClusterAgentNetworkPolicy(dda *datadoghqv1alpha1.DatadogAgent, name st
 		},
 	}
 	return policy
-}
-
-const (
-	defaultKSMCoreConfigMap = `
----
-cluster_check: true
-init_config:
-instances:
-  - prometheus_timeout: 30
-    min_collection_interval: 30
-    collectors:
-      - pods
-      - replicationcontrollers
-      - statefulsets
-      - nodes
-      - cronjobs
-      - jobs
-      - replicasets
-    label_joins:
-      kube_job_labels:
-          labels_to_match:
-            - job_name
-            - namespace
-          labels_to_get:
-            - label_service
-            - label_chart_name
-            - label_chart_version
-            - label_team
-            - label_app
-    labels_mapper:
-      label_service: service
-      label_chart_name: chart_name
-      label_chart_version: chart_version
-      label_team: team
-      label_app: app
-    telemetry: true
-  - prometheus_timeout: 30
-    min_collection_interval: 20
-    collectors:
-      - deployments
-      - configmaps
-      - services
-      - endpoints
-      - daemonsets
-      - horizontalpodautoscalers
-      - limitranges
-      - resourcequotas
-      - secrets
-      - namespaces
-      - persistentvolumeclaims
-      - persistentvolumes
-    label_joins:
-      kube_deployment_labels:
-          labels_to_match:
-            - deployment
-            - namespace
-          labels_to_get:
-            - label_service
-            - label_chart_name
-            - label_chart_version
-            - label_team
-            - label_app
-      kube_daemonset_labels:
-          labels_to_match:
-            - daemonset
-            - namespace
-          labels_to_get:
-            - label_service
-            - label_chart_name
-            - label_chart_version
-            - label_team
-            - label_app
-    labels_mapper:
-      label_service: service
-      label_chart_name: chart_name
-      label_chart_version: chart_version
-      label_team: team
-      label_app: app
-    telemetry: true
-`
-	ksmCoreCheckName = "kubernetes_state_core.yaml"
-)
-
-func buildKSMCoreConfigMap(dda *datadoghqv1alpha1.DatadogAgent) (*corev1.ConfigMap, error) {
-	configMap := &corev1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        datadoghqv1alpha1.GetKubeStateMetricsConfName(dda),
-			Namespace:   dda.Namespace,
-			Labels:      getDefaultLabels(dda, dda.Name, getAgentVersion(dda)),
-			Annotations: getDefaultAnnotations(dda),
-		},
-		Data: map[string]string{
-			ksmCoreCheckName: defaultKSMCoreConfigMap,
-		},
-	}
-	return configMap, nil
-}
-
-func (r *Reconciler) createKubeStateMetricsClusterRole(logger logr.Logger, dda *datadoghqv1alpha1.DatadogAgent, name, version string) (reconcile.Result, error) {
-	clusterRole := buildKubeStateMetricsCoreRBAC(dda, name, version)
-	logger.Info("createKubeStateMetricsClusterRole", "clusterRole.name", clusterRole.Name)
-	event := buildEventInfo(clusterRole.Name, clusterRole.Namespace, clusterRoleKind, datadog.CreationEvent)
-	r.recordEvent(dda, event)
-	return reconcile.Result{Requeue: true}, r.client.Create(context.TODO(), clusterRole)
-}
-
-// buildKubeStateMetricsCoreRBAC generates the cluster role required for the KSM informers to query
-// what is exposed as of the v2.0 https://github.com/kubernetes/kube-state-metrics/blob/release-2.0/examples/standard/cluster-role.yaml
-func buildKubeStateMetricsCoreRBAC(dda *datadoghqv1alpha1.DatadogAgent, name, version string) *rbacv1.ClusterRole {
-	clusterRole := &rbacv1.ClusterRole{
-		ObjectMeta: metav1.ObjectMeta{
-			Labels: getDefaultLabels(dda, name, version),
-			Name:   name,
-		},
-	}
-
-	rbacRules := []rbacv1.PolicyRule{
-		{
-			APIGroups: []string{datadoghqv1alpha1.CoreAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.ConfigMapsResource,
-				datadoghqv1alpha1.SecretsResource,
-				datadoghqv1alpha1.NodesResource,
-				datadoghqv1alpha1.PodsResource,
-				datadoghqv1alpha1.ServicesResource,
-				datadoghqv1alpha1.ResourceQuotasResource,
-				datadoghqv1alpha1.ReplicationControllersResource,
-				datadoghqv1alpha1.LimitRangesResource,
-				datadoghqv1alpha1.PersistentVolumeClaimsResource,
-				datadoghqv1alpha1.PersistentVolumesResource,
-				datadoghqv1alpha1.NamespaceResource,
-				datadoghqv1alpha1.EndpointsResource,
-			},
-		},
-		{
-			APIGroups: []string{datadoghqv1alpha1.ExtensionsAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.DaemonsetsResource,
-				datadoghqv1alpha1.DeploymentsResource,
-				datadoghqv1alpha1.ReplicasetsResource,
-			},
-		},
-		{
-			APIGroups: []string{datadoghqv1alpha1.AppsAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.StatefulsetsResource,
-				datadoghqv1alpha1.DaemonsetsResource,
-				datadoghqv1alpha1.DeploymentsResource,
-				datadoghqv1alpha1.ReplicasetsResource,
-			},
-		},
-		{
-			APIGroups: []string{datadoghqv1alpha1.BatchAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.CronjobsResource,
-				datadoghqv1alpha1.JobsResource,
-			},
-		},
-		{
-			APIGroups: []string{datadoghqv1alpha1.AutoscalingAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.HorizontalPodAutoscalersRecource,
-			},
-		},
-		{
-			APIGroups: []string{datadoghqv1alpha1.PolicyAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.PodDisruptionBudgetsResource,
-			},
-		},
-		{
-			APIGroups: []string{datadoghqv1alpha1.CertificatesAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.CertificatesSigningRequestsResource,
-			},
-		},
-		{
-			APIGroups: []string{datadoghqv1alpha1.StorageAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.StorageClassesResource,
-				datadoghqv1alpha1.VolumeAttachments,
-			},
-		},
-		{
-			APIGroups: []string{datadoghqv1alpha1.AdmissionAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.MutatingConfigResource,
-				datadoghqv1alpha1.ValidatingConfigResource,
-			},
-		},
-		{
-			APIGroups: []string{datadoghqv1alpha1.NetworkingAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.NetworkPolicyResource,
-				datadoghqv1alpha1.IngressesResource,
-			},
-		},
-		{
-			APIGroups: []string{datadoghqv1alpha1.CoordinationAPIGroup},
-			Verbs: []string{
-				datadoghqv1alpha1.ListVerb,
-				datadoghqv1alpha1.WatchVerb,
-			},
-			Resources: []string{
-				datadoghqv1alpha1.LeasesResource,
-			},
-		},
-	}
-
-	//commonVerbs := []string{
-	//	datadoghqv1alpha1.ListVerb,
-	//	datadoghqv1alpha1.WatchVerb,
-	//}
-	//for _, rule := range rbacRules {
-	//	rule.Verbs = append(rule.Verbs, commonVerbs...)
-	//}
-	clusterRole.Rules = rbacRules
-	return clusterRole
 }
